% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottle_code.R
\name{bottle_keywords}
\alias{bottle_keywords}
\title{Find all keywords for bottle}
\usage{
bottle_keywords(expr_str, crate_storage = FALSE)
}
\arguments{
\item{expr_str}{String with the code expression to be analysed}

\item{crate_storage}{}
}
\value{
Vector of keywords detected in the code
}
\description{
Look for all of the keywords from the code block and return all of the unique entries
}
