% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crate.R
\name{Crate}
\alias{Crate}
\title{Crate - Collection of bottles}
\usage{
Crate$bottle(id)({ 1 + 1 })
}
\description{
A crate is a collection of bottles which ensures that all of the bottles
share the same env. The aim is for cases where you generate multiple bottles with repeated.\cr\cr

A crate provides multiple interfaces

It is very important to consider that all of the bottles share the same namespace!
This means that if you have variables that overlab in the bottles, they will all have the same value.
Aim to store top level data in the object, and do any specific filtering within the bottle!
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Crate-new}{\code{Crate$new()}}
\item \href{#method-Crate-bottle}{\code{Crate$bottle()}}
\item \href{#method-Crate-get_bottle}{\code{Crate$get_bottle()}}
\item \href{#method-Crate-run_bottle}{\code{Crate$run_bottle()}}
\item \href{#method-Crate-print}{\code{Crate$print()}}
\item \href{#method-Crate-save}{\code{Crate$save()}}
\item \href{#method-Crate-add_to_env}{\code{Crate$add_to_env()}}
\item \href{#method-Crate-get_from_env}{\code{Crate$get_from_env()}}
\item \href{#method-Crate-clone}{\code{Crate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-new"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-new}{}}}
\subsection{Method \code{new()}}{
Constructor function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$new(filename = NULL, packages = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{packages}}{Add some packages that should be loaded with any bottle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Crate}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-bottle"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-bottle}{}}}
\subsection{Method \code{bottle()}}{
Bottle some code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$bottle(..., search_env = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Names for the id of the bottle}

\item{\code{search_env}}{Environment to search for variables and data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Function that takes in an expression to bottle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-get_bottle"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-get_bottle}{}}}
\subsection{Method \code{get_bottle()}}{
Get a bottle from the crate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$get_bottle(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Id of the bottle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Bottle if present otherwise NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-run_bottle"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-run_bottle}{}}}
\subsection{Method \code{run_bottle()}}{
Run bottle with a specific id from the crate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$run_bottle(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Id of the bottle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of the bottle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-print"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-print}{}}}
\subsection{Method \code{print()}}{
Print get an overview of
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Object}

\item{\code{...}}{Ignored parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-save"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-save}{}}}
\subsection{Method \code{save()}}{
Save the crate as an .rda file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$save(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{rda file that should store the data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-add_to_env"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-add_to_env}{}}}
\subsection{Method \code{add_to_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$add_to_env(var, env = parent.frame())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-get_from_env"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-get_from_env}{}}}
\subsection{Method \code{get_from_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$get_from_env(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
