% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottle_code.R
\name{ignore}
\alias{ignore}
\title{Mark a variable to be ignored by the keyword detection}
\usage{
ignore(var)
}
\arguments{
\item{var}{Variable to be marked}
}
\value{
return the var without any modifications
}
\description{
Can be used as a quick fix if unexpected values or packages are added to the bottle environment
}
