% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_bottle.R
\name{ls_bottle}
\alias{ls_bottle}
\title{List bottle environment}
\usage{
ls_bottle(bottle)
}
\arguments{
\item{bottle}{Input bottle to look up environment variables}
}
\description{
List all of the items available in the environment
Considers both the standalone bottle case and the bottles from crates with shared environments
}
