% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottle_code.R
\name{shared}
\alias{shared}
\title{Mark a variable as being shared in a bottle code}
\usage{
shared(var)
}
\arguments{
\item{var}{Variable to be marked}
}
\value{
return the var without any modifications
}
\description{
Helps the keyword detection to remove the keyword so it is not added to the variable again
}
