% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottle_code.R
\name{unshare}
\alias{unshare}
\title{Mark a variable as explicitly that is should not be shared}
\usage{
unshare(var)
}
\arguments{
\item{var}{Variable to be marked}
}
\value{
return the var without any modifications
}
\description{
Helps keyword detection to list a keyword to be
}
