% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottle_code.R
\name{bottle_keywords}
\alias{bottle_keywords}
\title{Find all keywords for bottle}
\usage{
bottle_keywords(expr_str)
}
\arguments{
\item{expr_str}{}
}
\description{
Look for all of the keywords from the code block and return all of the unique entries
}
