% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crate.R
\name{Crate}
\alias{Crate}
\title{Crate - Collection of bottles}
\usage{
Crate$bottle(id)({ 1 + 1 })
}
\description{
A crate is a collection of bottles which ensures that all of the bottles
share the same env. The aim is for cases where you generate multiple bottles with repeated.\cr\cr

A crate provides multiple interfaces

It is very important to consider that all of the bottles share the same namespace!
This means that if you have variables that overlab in the bottles, they will all have the same value.
Aim to store top level data in the object, and do any specific filtering within the bottle!
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Crate-new}{\code{Crate$new()}}
\item \href{#method-Crate-bottle}{\code{Crate$bottle()}}
\item \href{#method-Crate-get_bottle}{\code{Crate$get_bottle()}}
\item \href{#method-Crate-run_bottle}{\code{Crate$run_bottle()}}
\item \href{#method-Crate-print}{\code{Crate$print()}}
\item \href{#method-Crate-save}{\code{Crate$save()}}
\item \href{#method-Crate-add_data}{\code{Crate$add_data()}}
\item \href{#method-Crate-get_data}{\code{Crate$get_data()}}
\item \href{#method-Crate-rm_data}{\code{Crate$rm_data()}}
\item \href{#method-Crate-ls_data}{\code{Crate$ls_data()}}
\item \href{#method-Crate-ls_bottles}{\code{Crate$ls_bottles()}}
\item \href{#method-Crate-clone}{\code{Crate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-new"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-new}{}}}
\subsection{Method \code{new()}}{
Constructor function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$new(filename = NULL, packages = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Load the Crate from file. Overwrites all behaviour}

\item{\code{packages}}{Add some packages that should be loaded with any bottle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Crate}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-bottle"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-bottle}{}}}
\subsection{Method \code{bottle()}}{
Bottle some code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$bottle(..., search_env = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Names for the id of the bottle}

\item{\code{search_env}}{Environment to search for variables and data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Function that takes in an expression to bottle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-get_bottle"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-get_bottle}{}}}
\subsection{Method \code{get_bottle()}}{
Get a bottle from the crate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$get_bottle(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Id of the bottle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Bottle if present otherwise NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-run_bottle"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-run_bottle}{}}}
\subsection{Method \code{run_bottle()}}{
Run bottle with a specific id from the crate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$run_bottle(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Id of the bottle}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of the bottle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-print"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-print}{}}}
\subsection{Method \code{print()}}{
Print get an overview of
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Object}

\item{\code{...}}{Ignored parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-save"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-save}{}}}
\subsection{Method \code{save()}}{
Save the crate as an .rda file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$save(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{rda file that should store the data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-add_data}{}}}
\subsection{Method \code{add_data()}}{
Add data into the shared environment. Bottles can use the shared() function to avoid
adding the data into it again.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$add_data(var, env = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Variable that should be added to the data}

\item{\code{env}}{Environment the variable is from. Default is the caller environment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-get_data}{}}}
\subsection{Method \code{get_data()}}{
Retreive a variable from the share data environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$get_data(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Variable in the shared data environment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The data of the variable or NULL if it does not exist
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-rm_data"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-rm_data}{}}}
\subsection{Method \code{rm_data()}}{
Remove a variable from the the shared data environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$rm_data(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Variable in the shared data environment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-ls_data"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-ls_data}{}}}
\subsection{Method \code{ls_data()}}{
List all of the names in the shared data environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$ls_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Names of items in the share data environment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-ls_bottles"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-ls_bottles}{}}}
\subsection{Method \code{ls_bottles()}}{
Get the names of each of the bottles stored in the crate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$ls_bottles()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of input names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Crate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
